+++ js/ngl/buffer.js
@@ -1178,7 +1178,8 @@ NGL.SphereImpostorBuffer = function( position, color, radius, pickingColor, para
     NGL.QuadBuffer.call( this, params );
 
     this.addUniforms( {
-        "projectionMatrixInverse": { value: new THREE.Matrix4() }
+        "projectionMatrixInverse": { value: new THREE.Matrix4() },
+        "ortho": { value: 0.0 },
     } );
 
     this.addAttributes( {
@@ -1239,6 +1240,7 @@ NGL.CylinderImpostorBuffer = function( from, to, color, color2, radius, pickingC
     this.addUniforms( {
         "modelViewMatrixInverse": modelViewMatrixInverse,
         "shift": { value: this.shift },
+        "ortho": { value: 0.0 },
     } );
 
     this.addAttributes( {