+++ js/ngl/stage.js
@@ -52,6 +52,7 @@ NGL.Stage = function( eid, params ){
         fogNear: 50,
         fogFar: 100,
         cameraFov: 40,
+        cameraType: "perspective",
         lightColor: 0xdddddd,
         lightIntensity: 1.0,
         ambientColor: 0xdddddd,
@@ -107,6 +108,9 @@ NGL.Stage.prototype = {
         fogFar: {
             type: "range", step: 1, max: 100, min: 0
         },
+        cameraType: {
+            type: "select", options: { "perspective": "perspective", "orthographic": "orthographic" }
+        },
         cameraFov: {
             type: "range", step: 1, max: 120, min: 15
         },
@@ -152,7 +156,7 @@ NGL.Stage.prototype = {
         if( p.panSpeed !== undefined ) controls.panSpeed = p.panSpeed;
         viewer.setClip( p.clipNear, p.clipFar, p.clipDist );
         viewer.setFog( undefined, p.fogNear, p.fogFar );
-        viewer.setCamera( undefined, p.cameraFov );
+        viewer.setCamera( p.cameraType, p.cameraFov );
         viewer.setSampling( p.sampleLevel );
         viewer.setBackground( p.backgroundColor );
         viewer.setLight(
@@ -1232,6 +1236,8 @@ NGL.StructureComponent.prototype = NGL.createObject(
     },
 
     centerView: function( zoom, sele ){
+
+        zoom = zoom !== undefined ? zoom : true;
 
         var center;